Moves batch related test files
